requirements.txt
pillow>=10.3.0
requests>=2.32.0
ttkbootstrap>=1.10.1
